refactor(scheduleDisplay): 优化类型定义和文档注释
